pyproject.toml
[tool.poetry]
name = "Cat_or_dog_breed"
version = "0.1.0"
description = ""
authors = ["Mustapha Ayadi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.3"
pillow = "^9.5.0"
matplotlib = "^3.7.1"
pyqt5 = "^5.15.9"
pandas = "^2.0.1"
torch = "^2.0.1"
torchvision = "^0.15.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
